# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta API Json Rest actua como proxy SOAP a los servicios web de ARCA.",
        "title": "API proxy a los webservices de ARCA",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Sebastian Hogas",
            "email": "sehogas@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/coemcons/Dummy": {
            "get": {
                "description": "Visualizar el estado del servicio web, del servicio de autenticación y de la base de datos de ARCA",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consultas de Comunicación de Embarque"
                ],
                "summary": "Muestra el estado del servicio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wscoemcons.ResultadoEjecucionOfDummyOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/coemcons/ObtenerConsultaEstadosCOEM": {
            "get": {
                "description": "Obtiene una lista de comunicaciones de embarque comprendidas en la cabecera informada.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consultas de Comunicación de Embarque"
                ],
                "summary": "Obtener Consulta Estados COEM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Identificador de la caratula",
                        "name": "identificadorCabecera",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wscoemcons.ResultadoEjecucionOfResultadoEstadoProceso"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/coemcons/ObtenerConsultaNoAbordo": {
            "get": {
                "description": "Obtiene una lista de aquellas comunicaciones de embarque relacionadas con la carátula de referencia para las cuales se ha definido el no abordaje parcial o total de su contenido, el cual puede tratarse de bultos sueltos, mercadería a granel, contenedores de carga o cont¢nedores vacíos.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consultas de Comunicación de Embarque"
                ],
                "summary": "Obtener Consulta No Abordo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Identificador de la caratula",
                        "name": "identificadorCabecera",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wscoemcons.ResultadoEjecucionOfResultadoNoAbordoProceso"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/coemcons/ObtenerConsultaSolicitudes": {
            "get": {
                "description": "Obtiene la lista de solicitudes de comunicaciones asociadas a la cabecera informada. Número y tipo de solicitud, estado y fecha.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consultas de Comunicación de Embarque"
                ],
                "summary": "Obtener Consulta de Solicitudes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Identificador de la caratula",
                        "name": "identificadorCabecera",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wscoemcons.ResultadoEjecucionOfResultadoSolicitudProceso"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "wscoemcons.ArrayOfConsultaEstadoCOEM": {
            "type": "object",
            "properties": {
                "ConsultaEstadoCOEM": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wscoemcons.ConsultaEstadoCOEM"
                    }
                }
            }
        },
        "wscoemcons.ArrayOfConsultaNoAbordo": {
            "type": "object",
            "properties": {
                "ConsultaNoAbordo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wscoemcons.ConsultaNoAbordo"
                    }
                }
            }
        },
        "wscoemcons.ArrayOfConsultaSolicitudes": {
            "type": "object",
            "properties": {
                "ConsultaSolicitudes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wscoemcons.ConsultaSolicitudes"
                    }
                }
            }
        },
        "wscoemcons.ArrayOfErrorEjecucion": {
            "type": "object",
            "properties": {
                "ErrorEjecucion": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wscoemcons.ErrorEjecucion"
                    }
                }
            }
        },
        "wscoemcons.ConsultaEstadoCOEM": {
            "type": "object",
            "properties": {
                "CODE": {
                    "type": "string"
                },
                "CuitAlta": {
                    "type": "string"
                },
                "Estado": {
                    "type": "string"
                },
                "FechaEstado": {
                    "type": "string"
                },
                "IdentificadorCOEM": {
                    "type": "string"
                },
                "Motivo": {
                    "type": "string"
                }
            }
        },
        "wscoemcons.ConsultaNoAbordo": {
            "type": "object",
            "properties": {
                "Contenedor": {
                    "type": "string"
                },
                "Cuit": {
                    "type": "string"
                },
                "DescripcionMotivo": {
                    "type": "string"
                },
                "Destinacion": {
                    "type": "string"
                },
                "FechaNoAbordo": {
                    "type": "string"
                },
                "IdentificadorCACE": {
                    "type": "string"
                },
                "IdentificadorCOEM": {
                    "type": "string"
                },
                "MotivoNoAbordo": {
                    "type": "string"
                },
                "Tipo": {
                    "type": "string"
                },
                "TipoNoAbordo": {
                    "type": "string"
                }
            }
        },
        "wscoemcons.ConsultaSolicitudes": {
            "type": "object",
            "properties": {
                "Estado": {
                    "type": "string"
                },
                "FechaEstado": {
                    "type": "string"
                },
                "IdentificadorCACE": {
                    "type": "string"
                },
                "IdentificadorCOEM": {
                    "type": "string"
                },
                "NumeroSolicitud": {
                    "type": "string"
                },
                "TipoSolicitud": {
                    "type": "string"
                }
            }
        },
        "wscoemcons.DummyOutput": {
            "type": "object",
            "properties": {
                "AppServer": {
                    "type": "string"
                },
                "AuthServer": {
                    "type": "string"
                },
                "DbServer": {
                    "type": "string"
                }
            }
        },
        "wscoemcons.ErrorEjecucion": {
            "type": "object",
            "properties": {
                "Codigo": {
                    "type": "string"
                },
                "Descripcion": {
                    "type": "string"
                }
            }
        },
        "wscoemcons.ResultadoEjecucionOfDummyOutput": {
            "type": "object",
            "properties": {
                "Errores": {
                    "$ref": "#/definitions/wscoemcons.ArrayOfErrorEjecucion"
                },
                "Resultado": {
                    "$ref": "#/definitions/wscoemcons.DummyOutput"
                },
                "Server": {
                    "type": "string"
                },
                "TimeStamp": {
                    "type": "string"
                }
            }
        },
        "wscoemcons.ResultadoEjecucionOfResultadoEstadoProceso": {
            "type": "object",
            "properties": {
                "Errores": {
                    "$ref": "#/definitions/wscoemcons.ArrayOfErrorEjecucion"
                },
                "Resultado": {
                    "$ref": "#/definitions/wscoemcons.ResultadoEstadoProceso"
                },
                "Server": {
                    "type": "string"
                },
                "TimeStamp": {
                    "type": "string"
                }
            }
        },
        "wscoemcons.ResultadoEjecucionOfResultadoNoAbordoProceso": {
            "type": "object",
            "properties": {
                "Errores": {
                    "$ref": "#/definitions/wscoemcons.ArrayOfErrorEjecucion"
                },
                "Resultado": {
                    "$ref": "#/definitions/wscoemcons.ResultadoNoAbordoProceso"
                },
                "Server": {
                    "type": "string"
                },
                "TimeStamp": {
                    "type": "string"
                }
            }
        },
        "wscoemcons.ResultadoEjecucionOfResultadoSolicitudProceso": {
            "type": "object",
            "properties": {
                "Errores": {
                    "$ref": "#/definitions/wscoemcons.ArrayOfErrorEjecucion"
                },
                "Resultado": {
                    "$ref": "#/definitions/wscoemcons.ResultadoSolicitudProceso"
                },
                "Server": {
                    "type": "string"
                },
                "TimeStamp": {
                    "type": "string"
                }
            }
        },
        "wscoemcons.ResultadoEstadoProceso": {
            "type": "object",
            "properties": {
                "InformacionAdicional": {
                    "type": "string"
                },
                "Listado": {
                    "$ref": "#/definitions/wscoemcons.ArrayOfConsultaEstadoCOEM"
                }
            }
        },
        "wscoemcons.ResultadoNoAbordoProceso": {
            "type": "object",
            "properties": {
                "InformacionAdicional": {
                    "type": "string"
                },
                "Listado": {
                    "$ref": "#/definitions/wscoemcons.ArrayOfConsultaNoAbordo"
                }
            }
        },
        "wscoemcons.ResultadoSolicitudProceso": {
            "type": "object",
            "properties": {
                "InformacionAdicional": {
                    "type": "string"
                },
                "Listado": {
                    "$ref": "#/definitions/wscoemcons.ArrayOfConsultaSolicitudes"
                }
            }
        },
...
    }
}